Feat/sd 48 process order product delivery

See merge request sea-labs-id/batch-01/stage-02/sea-deals/seadeals-backend!30

+++ model/product_detail.go
@@ -4,14 +4,14 @@ import "gorm.io/gorm"
 
 type ProductDetail struct {
 	gorm.Model      `json:"-"`
-	ID              uint    `json:"id" gorm:"primaryKey"`
-	ProductID       uint    `json:"product_id"`
-	Description     string  `json:"description"`
-	VideoURL        string  `json:"video_url"`
-	IsHazardous     bool    `json:"is_hazardous"`
-	ConditionStatus string  `json:"condition_status"`
-	Length          float64 `json:"length"`
-	Width           float64 `json:"width"`
-	Height          float64 `json:"height"`
-	Weight          float64 `json:"weight"`
+	ID              uint   `json:"id" gorm:"primaryKey"`
+	ProductID       uint   `json:"product_id"`
+	Description     string `json:"description"`
+	VideoURL        string `json:"video_url"`
+	IsHazardous     bool   `json:"is_hazardous"`
+	ConditionStatus string `json:"condition_status"`
+	Length          int    `json:"length"`
+	Width           int    `json:"width"`
+	Height          int    `json:"height"`
+	Weight          int    `json:"weight"`
 }
